# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": "Api do gerenciamento de times. Usuário Admin para testes [ Email: admin@example.com | Senha: admin123 ] "
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
...
    "/teams/create": {
      "post": {
        "tags": ["Teams"],
        "description": "Cria um time.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "team-name": {
                  "example": "team name"
                },
                "description": {
                  "example": "team description"
                },
                "teamId": {
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Time criado com sucesso!"
          },
          "400": {
            "description": "Dados inválidos."
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "teams/{teamId}/member/{userId}": {
      "post": {
        "tags": ["Teams"],
        "description": "Adicionar um membro a um time.",
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "teams/{teamId}/member/{id}": {
      "delete": {
        "tags": ["Teams"],
        "description": "Remove um membro do time.",
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do time do qual o membro será removido."
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do usuário a ser removido do time."
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário removido do time com sucesso!"
          },
          "400": {
            "description": "Dados incorretos, tente novamente."
          },
          "403": {
            "description": "Você não tem permissão para remover membros deste time."
          },
          "404": {
            "description": "Time ou usuário não encontrado."
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "teams/read": {
      "get": {
        "tags": ["Teams"],
        "description": "Busca times com paginação.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "example": 10
            },
            "description": "Máximo de times exibidos por página."
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "Número da página a ser retornada."
          }
        ],
        "responses": {
          "200": {
            "description": "Times encontrados!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Times encontrados!"
                    },
                    "teams": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Team"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados de paginação inválidos."
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "teams/my-teams": {
      "get": {
        "tags": ["Teams"],
        "description": "Busca os times dos quais o usuário é membro ou líder.",
        "responses": {
          "200": {
            "description": "Times pertencentes encontrados.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Times pertencentes encontrados."
                    },
                    "teams": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Team"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "teams/update/{id}": {
      "put": {
        "tags": ["Teams"],
        "description": "Atualiza os dados de um time.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do time a ser atualizado."
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Novo Nome do Time"
                },
                "description": {
                  "type": "string",
                  "example": "Nova descrição do time."
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados do time atualizados com sucesso!"
          },
          "400": {
            "description": "Dados inválidos."
          },
          "403": {
            "description": "Você não tem permissão para atualizar este time."
          },
          "404": {
            "description": "Time não encontrado."
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "teams/delete/{id}": {
      "delete": {
        "tags": ["Teams"],
        "description": "Deleta um time.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do time a ser deletado."
          }
        ],
        "responses": {
          "204": {
            "description": "Time deletado com sucesso!"
          },
          "400": {
            "description": "Dados inválidos."
          },
          "403": {
            "description": "Você não tem permissão para deletar este time."
          },
          "404": {
            "description": "Time não encontrado."
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}